remove one layer of abstraction
